Break a list into chunks of size N
The yield keyword enables a function to comeback where it left off when it is called again. The yield keyword helps a function to remember its state. The yield enables a function to suspend and resume while it turns in a value at the time of the suspension of the execution.